refactor(bartworks): isolate platinum recipes
